OpenWRT: WebGui: many bugfixes

+++ openwrt/qaul.net/qaul/bin/logout.sh
@@ -6,6 +6,9 @@ IFS="&="
 set -- $HTTP_COOKIE
 MD5STRING=$2
 
-rm /qaul/auth/$MD5STRING
+if [ "$MD5STRING" != "" ]
+then
+	rm /qaul/auth/$MD5STRING
+fi
 
 exit 0
